Started refactoring the configuration gear.
