+++ buildkite/config/cli.go
@@ -1,4 +1,4 @@
-package buildkite
+package config
 
 import (
 	"github.com/buildkite/agent/buildkite/logger"
@@ -8,6 +8,9 @@ import (
 type CLI struct {
 	// The cli context for this command
 	Context *cli.Context
+}
+
+func (c CLI) Configure(obj interface{}) {
 }
 
 func (c CLI) Setup() CLI {
